Add mode to invert axis
This patch adds the possibilty to mirror/invert the document on one or both
of the axis. This is helpful if the plotter has a different origin of the
coordinate system.

This patch also renames the existing "flip" option, which multiplies the value
of an axis with -1 to "negate" to make it more clear what it does.