Move plotter configuration to configuration file
Removes the possibility to provide the configuration strings
via the command line and moves the config over to the vpype
configuration.

This changes the command line option to -p or --profile. It also
allows to configure a default profile using the `default_profile`
option in the [gcode] section of the vypye configuration. The cmd
line parameter is not required anymore. An error is thrown if
there is no default set and no profile was provided explicitly.