Better handle the output file
Using click.File we can be sure that the file is opened.
Additionally, the actual opening and closing of the file
can be left to click and one line of indent can be removed.

Also as a nice bonus, this allows to output the gcode to
stdout by providing "-" as output file on the command line
similar to typical unix tools.