Added support for property substitution in profile definition.
Layer and global property may now be used in profile definition (e.g. `{vp_color}`). To avoid error when they are missing, default value may be provided by the profile (`default_values` key) or in the CLI (`--default` option). The latter is not recommended for non-string value as number-specific formatting (`{...:d}`, etc.) would break.

The following change were also introduced:
- bumped vpype to 1.10
- updated list of compatible python version
- removed `flush()` and `close()` on file (done automatically by click)
- print `info` to stderr
- applied black formatting on gwrite.py

Some of the changes above improves the use of `-` as path to output to stdout instead of in a file.